no extra type file
fix pyproject groups

fix lockfile

add pre-commit ci

actually run ruff and fix things
