+++ pyproject.toml
@@ -24,15 +24,16 @@ build-backend = "setuptools.build_meta"
 [tool.setuptools_scm]
 write_to = "src/fastmcp/_version.py"
 
-[dependency-groups]
+[project.optional-dependencies]
 dev = [
     "copychat>=0.5.2",
     "ipython>=8.12.3",
     "pdbpp>=0.10.3",
-    "pre-commit>=3.10.1",
+    "pre-commit",
     "pytest-xdist>=3.6.1",
     "pytest>=8.3.3",
     "pytest-asyncio>=0.23.5",
+    "ruff",
 ]
 
 [tool.pytest.ini_options]
